chore: Piece 종류 별 패키지 분리
