+++ src/main/java/chess/domain/piece/common/Queen.java
@@ -1,19 +1,26 @@
-package chess.domain.piece;
+package chess.domain.piece.common;
 
 import static chess.domain.piece.Direction.E;
 import static chess.domain.piece.Direction.N;
+import static chess.domain.piece.Direction.NE;
+import static chess.domain.piece.Direction.NW;
 import static chess.domain.piece.Direction.S;
+import static chess.domain.piece.Direction.SE;
+import static chess.domain.piece.Direction.SW;
 import static chess.domain.piece.Direction.W;
 
 import chess.domain.board.Position;
+import chess.domain.piece.Color;
+import chess.domain.piece.Direction;
+import chess.domain.piece.PieceType;
 import java.util.List;
 
-public class Rook extends CommonPiece {
+public class Queen extends CommonPiece {
 
-    private static final List<Direction> POSSIBLE_DIRECTIONS = List.of(N, S, W, E);
+    private static final List<Direction> POSSIBLE_DIRECTIONS = List.of(N, S, W, E, NE, SE, SW, NW);
 
-    public Rook(final Color color) {
-        super(PieceType.ROOK, color);
+    public Queen(final Color color) {
+        super(PieceType.QUEEN, color);
     }
 
     @Override
