+++ src/main/java/chess/domain/piece/factory/PieceFactory.java
@@ -7,15 +7,15 @@ import static chess.domain.piece.PieceType.PAWN;
 import static chess.domain.piece.PieceType.QUEEN;
 import static chess.domain.piece.PieceType.ROOK;
 
-import chess.domain.piece.Bishop;
 import chess.domain.piece.Color;
-import chess.domain.piece.King;
-import chess.domain.piece.Knight;
 import chess.domain.piece.Pawn;
 import chess.domain.piece.Piece;
 import chess.domain.piece.PieceType;
-import chess.domain.piece.Queen;
-import chess.domain.piece.Rook;
+import chess.domain.piece.common.Bishop;
+import chess.domain.piece.common.Queen;
+import chess.domain.piece.common.Rook;
+import chess.domain.piece.ranged.King;
+import chess.domain.piece.ranged.Knight;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.function.Function;
