+++ src/test/java/chess/domain/StatusTest.java
@@ -6,10 +6,10 @@ import chess.domain.board.Board;
 import chess.domain.board.Position;
 import chess.domain.board.strategy.CreateMockBoardStrategy;
 import chess.domain.piece.Color;
-import chess.domain.piece.King;
 import chess.domain.piece.Pawn;
 import chess.domain.piece.Piece;
-import chess.domain.piece.Queen;
+import chess.domain.piece.common.Queen;
+import chess.domain.piece.ranged.King;
 import java.util.HashMap;
 import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
