+++ src/test/java/chess/domain/board/BoardTest.java
@@ -6,15 +6,15 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 import chess.domain.CachedPosition;
 import chess.domain.board.strategy.CreateCompleteBoardStrategy;
 import chess.domain.board.strategy.CreateMockBoardStrategy;
-import chess.domain.piece.Bishop;
 import chess.domain.piece.Color;
-import chess.domain.piece.King;
-import chess.domain.piece.Knight;
 import chess.domain.piece.Pawn;
 import chess.domain.piece.Piece;
 import chess.domain.piece.PieceType;
-import chess.domain.piece.Queen;
-import chess.domain.piece.Rook;
+import chess.domain.piece.common.Bishop;
+import chess.domain.piece.common.Queen;
+import chess.domain.piece.common.Rook;
+import chess.domain.piece.ranged.King;
+import chess.domain.piece.ranged.Knight;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.stream.Stream;
