+++ src/test/java/chess/domain/piece/common/BishopTest.java
@@ -1,10 +1,11 @@
-package chess.domain.piece;
+package chess.domain.piece.common;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import chess.domain.CachedPosition;
 import chess.domain.board.Position;
+import chess.domain.piece.Color;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
@@ -25,18 +26,6 @@ class BishopTest {
             Bishop bishop = new Bishop(Color.BLACK);
 
             assertThatThrownBy(() -> bishop.findValidDirection(current, invalidTarget));
-        }
-
-        @DisplayName("유효하면 방향 객체를 반환한다.")
-        @Test
-        void valid_Direction() {
-            Position current = CachedPosition.a1;
-            Position target = CachedPosition.c3;
-            Bishop bishop = new Bishop(Color.BLACK);
-
-            Direction actual = bishop.findValidDirection(current, target);
-
-            assertThat(actual).isEqualTo(Direction.NE);
         }
     }
