+++ src/test/java/chess/domain/piece/ranged/KnightTest.java
@@ -1,10 +1,12 @@
-package chess.domain.piece;
+package chess.domain.piece.ranged;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import chess.domain.CachedPosition;
 import chess.domain.board.Position;
+import chess.domain.piece.Color;
+import chess.domain.piece.Direction;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
